Stop the phase of the note resetting so much
A clipping sound previously could be caused by the note phase being
reset if the note is being re-articulated, but the first articulation is
still sounding. Now, any note that is sounding will not have its phase
reset.

My ideal vision for phase resetting is that it can be used to ensure
that every note is in phase with every other note - sort of a 'global
phase' - but that is in the far future. This just fixes an immediate
issue.